+++ src/ruleset/rule.rs
@@ -4,14 +4,20 @@ use crate::{expr::Expr, function::FunctionContext, value::Value, Result};
 #[derive(Debug, Clone, PartialEq)]
 pub struct Rule {
     pub(super) name: String,
+    description: Option<String>,
     expr: Expr,
 }
 
 impl Rule {
     /// Construct a new rule from a name and an expression
-    pub fn new(name: impl Into<String>, expr: Expr) -> Self {
+    pub fn new(
+        name: impl Into<String>,
+        description: impl Into<Option<String>>,
+        expr: Expr,
+    ) -> Self {
         Self {
             name: name.into(),
+            description: description.into(),
             expr,
         }
     }
@@ -28,6 +34,11 @@ impl Rule {
     /// Return the name of the rule
     pub fn name(&self) -> &str {
         &self.name
+    }
+
+    /// Return an optional description of the rule
+    pub fn description(&self) -> Option<&str> {
+        self.description.as_deref()
     }
 }